initial

+++ webhooks.go
@@ -1,7 +1,7 @@
 package dusupay
 
 type CollectionWebhook struct {
-	ID                int     `json:"id"`
+	ID                int64   `json:"id"`
 	RequestAmount     float64 `json:"request_amount"`
 	RequestCurrency   string  `json:"request_currency"`
 	AccountAmount     float64 `json:"account_amount"`
@@ -21,7 +21,7 @@ type CollectionWebhook struct {
 }
 
 type PayoutWebhook struct {
-	ID                int     `json:"id"`
+	ID                int64   `json:"id"`
 	RequestAmount     float64 `json:"request_amount"`
 	RequestCurrency   string  `json:"request_currency"`
 	AccountAmount     float64 `json:"account_amount"`
@@ -39,16 +39,16 @@ type PayoutWebhook struct {
 	InstitutionName   string  `json:"institution_name"`
 }
 type RefundWebhook struct {
-	ID                  int    `json:"id"`
-	RefundAmount        int    `json:"refund_amount"`
-	RefundCurrency      string `json:"refund_currency"`
-	TransactionFee      int    `json:"transaction_fee"`
-	TotalDebit          int    `json:"total_debit"`
-	ProviderID          string `json:"provider_id"`
-	CollectionReference string `json:"collection_reference"`
-	InternalReference   string `json:"internal_reference"`
-	TransactionType     string `json:"transaction_type"`
-	TransactionStatus   string `json:"transaction_status"`
-	AccountNumber       string `json:"account_number"`
-	Message             string `json:"message"`
+	ID                  int64   `json:"id"`
+	RefundAmount        float64 `json:"refund_amount"`
+	RefundCurrency      string  `json:"refund_currency"`
+	TransactionFee      float64 `json:"transaction_fee"`
+	TotalDebit          float64 `json:"total_debit"`
+	ProviderID          string  `json:"provider_id"`
+	CollectionReference string  `json:"collection_reference"`
+	InternalReference   string  `json:"internal_reference"`
+	TransactionType     string  `json:"transaction_type"`
+	TransactionStatus   string  `json:"transaction_status"`
+	AccountNumber       string  `json:"account_number"`
+	Message             string  `json:"message"`
 }
